Moved cumulus image proxy helper to base project
