+++ controllers/image.js
@@ -3,7 +3,7 @@
 const assert = require('assert');
 
 const config = require('../collections-online/lib/config');
-const proxy = require('../collections-online-cumulus/proxy');
+const proxy = require('./image__proxy');
 
 assert.ok(config.cip, 'Missing config.cip');
 assert.ok(config.cip.proxy, 'Missing config.cip.proxy');
